+++ src/pages/BrowsePage.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { MovieList } from '../components';
-import MovieFeed from '../components/MovieFeed';
+import MovieFeedFlip from '../components/MovieFeedFlip';
 import '../css/BrowsePage.css';
 
 const BrowsePage = () => (
   <div>
     {/* <MovieList showGenreFilter results={[{}]} /> */}
-    <MovieFeed />
+    <MovieFeedFlip />
   </div>
 );
